AVR: Fix warnings from pointer casts and prototype
There was broken logic around serial initialization function
prototype.  Pointers were also being cast directly to uint32
without first being cast to uintptr.

+++ arch/avr/src/avr/up_initialstate.c
@@ -85,12 +85,12 @@ void up_initial_state(struct tcb_s *tcb)
   /* Save the task entry point */
 
 #if !defined(REG_PC2)
-  xcp->regs[REG_PC0]   = (uint8_t)((uint16_t)tcb->start >> 8);
-  xcp->regs[REG_PC1]   = (uint8_t)((uint16_t)tcb->start & 0xff);
+  xcp->regs[REG_PC0]   = (uint8_t)((uintptr_t)tcb->start >> 8);
+  xcp->regs[REG_PC1]   = (uint8_t)((uintptr_t)tcb->start & 0xff);
 #else
-  xcp->regs[REG_PC0]   = (uint8_t)((uint32_t)tcb->start >> 16);
-  xcp->regs[REG_PC1]   = (uint8_t)((uint32_t)tcb->start >> 8);
-  xcp->regs[REG_PC2]   = (uint8_t)((uint32_t)tcb->start & 0xff);
+  xcp->regs[REG_PC0]   = (uint8_t)((uint32_t)(uintptr_t)tcb->start >> 16);
+  xcp->regs[REG_PC1]   = (uint8_t)((uintptr_t)tcb->start >> 8);
+  xcp->regs[REG_PC2]   = (uint8_t)((uintptr_t)tcb->start & 0xff);
 #endif
 
   /* Enable or disable interrupts, based on user configuration */
